chore(system): move resolveResponsivePropStyles to its file
